Added utils directory
