+++ app/src/utils/passwordStrength.py
